refactor: move registerComponentFragment to Base
