+++ tools/base/tsconfig.json
@@ -5,6 +5,5 @@
         "rootDir": "./"
     },
     "include": ["*.ts", "src/*.ts"],
-    "exclude": ["test/*.ts", "stories/*.ts"],
-    "references": [{ "path": "../theme" }]
+    "exclude": ["test/*.ts", "stories/*.ts"]
 }
