+++ tools/theme/package.json
@@ -133,8 +133,8 @@
         "lit-html"
     ],
     "dependencies": {
-        "@spectrum-web-components/styles": "^0.40.3",
-        "lit": "^2.5.0"
+        "@spectrum-web-components/base": "^0.40.4",
+        "@spectrum-web-components/styles": "^0.40.3"
     },
     "types": "./src/index.d.ts",
     "customElements": "custom-elements.json",
